refactor write api
