+++ core/src/types/blocking_read/std_bytes_iterator.rs
